// src/types/enum/entity.ts
export enum CurveArithEnum {
  ARITH_NONE = 0,
  ARITH_ADD = 1,
  ARITH_MULTI = 2
}

export enum LifeStateEnum {
  LIFE_NONE = 0,
  LIFE_ALIVE = 1,
  LIFE_DEAD = 2,
  LIFE_REVIVE = 3
}

export enum MotionStateEnum {
  MOTION_NONE = 0,
  MOTION_RESET = 1,
  MOTION_STANDBY = 2,
  MOTION_STANDBY_MOVE = 3,
  MOTION_WALK = 4,
  MOTION_RUN = 5,
  MOTION_DASH = 6,
  MOTION_CLIMB = 7,
  MOTION_CLIMB_JUMP = 8,
  MOTION_STANDBY_TO_CLIMB = 9,
  MOTION_FIGHT = 10,
  MOTION_JUMP = 11,
  MOTION_DROP = 12,
  MOTION_FLY = 13,
  MOTION_SWIM_MOVE = 14,
  MOTION_SWIM_IDLE = 15,
  MOTION_SWIM_DASH = 16,
  MOTION_SWIM_JUMP = 17,
  MOTION_SLIP = 18,
  MOTION_GO_UPSTAIRS = 19,
  MOTION_FALL_ON_GROUND = 20,
  MOTION_JUMP_UP_WALL_FOR_STANDBY = 21,
  MOTION_JUMP_OFF_WALL = 22,
  MOTION_POWERED_FLY = 23,
  MOTION_LADDER_IDLE = 24,
  MOTION_LADDER_MOVE = 25,
  MOTION_LADDER_SLIP = 26,
  MOTION_STANDBY_TO_LADDER = 27,
  MOTION_LADDER_TO_STANDBY = 28,
  MOTION_DANGER_STANDBY = 29,
  MOTION_DANGER_STANDBY_MOVE = 30,
  MOTION_DANGER_WALK = 31,
  MOTION_DANGER_RUN = 32,
  MOTION_DANGER_DASH = 33,
  MOTION_CROUCH_IDLE = 34,
  MOTION_CROUCH_MOVE = 35,
  MOTION_CROUCH_ROLL = 36,
  MOTION_NOTIFY = 37,
  MOTION_LAND_SPEED = 38,
  MOTION_MOVE_FAIL_ACK = 39,
  MOTION_WATERFALL = 40,
  MOTION_DASH_BEFORE_SHAKE = 41,
  MOTION_SIT_IDLE = 42,
  MOTION_FORCE_SET_POS = 43,
  MOTION_QUEST_FORCE_DRAG = 44,
  MOTION_FOLLOW_ROUTE = 45,
  MOTION_SKIFF_BOARDING = 46,
  MOTION_SKIFF_NORMAL = 47,
  MOTION_SKIFF_DASH = 48,
  MOTION_SKIFF_POWERED_DASH = 49,
  MOTION_DESTROY_VEHICLE = 50,
  MOTION_FLY_IDLE = 51,
  MOTION_FLY_SLOW = 52,
  MOTION_FLY_FAST = 53,
  MOTION_NUM = 54
}

export enum ProtEntityTypeEnum {
  PROT_ENTITY_NONE = 0,
  PROT_ENTITY_AVATAR = 1,
  PROT_ENTITY_MONSTER = 2,
  PROT_ENTITY_NPC = 3,
  PROT_ENTITY_GADGET = 4,
  PROT_ENTITY_REGION = 5,
  PROT_ENTITY_WEAPON = 6,
  PROT_ENTITY_WEATHER = 7,
  PROT_ENTITY_SCENE = 8,
  PROT_ENTITY_TEAM = 9,
  PROT_ENTITY_MASSIVE_ENTITY = 10,
  PROT_ENTITY_MP_LEVEL = 11,
  PROT_ENTITY_PLAY_TEAM_ENTITY = 12,
  PROT_ENTITY_EYE_POINT = 13,
  PROT_ENTITY_MAX = 14
}

export enum VisionTypeEnum {
  VISION_NONE = 0,
  VISION_MEET = 1,
  VISION_REBORN = 2,
  VISION_REPLACE = 3,
  VISION_WAYPOINT_REBORN = 4,
  VISION_MISS = 5,
  VISION_DIE = 6,
  VISION_GATHER_ESCAPE = 7,
  VISION_REFRESH = 8,
  VISION_TRANSPORT = 9,
  VISION_REPLACE_DIE = 10,
  VISION_REPLACE_NO_NOTIFY = 11,
  VISION_BORN = 12,
  VISION_PICKUP = 13,
  VISION_REMOVE = 14,
  VISION_CHANGE_COSTUME = 15,
  VISION_FISH_REFRESH = 16,
  VISION_FISH_BIG_SHOCK = 17,
  VISION_FISH_QTE_SUCC = 18
}